Update Docker support

+++ docker-entrypoint.sh
@@ -3,19 +3,6 @@ set -e
 
 cd $(dirname $0)
 
-if [ -d build ]; then
-  rm -rf build
-fi
+[ -d build ] && rm -rf build
 
-mkdir -p build
-cd build
-cmake ..
-make -j6
-
-ARGS=""
-if [ -f ../debug-public-key.pem ]; then
-  ARGS="${ARGS} --debug-auth-pubkey=../debug-public-key.pem"
-fi
-twilio microvisor:apps:bundle -l debug ${ARGS} \
-  Demo/gpio_toggle_demo.bin \
-  Demo/gpio_toggle_demo.zip
+./deploy.sh --genkeys --log
